Cast format argument to expected type
Please GCC conversion warning:

    src/sp_upload_validation.c: In function 'sp_rfc1867_callback':
    src/sp_utils.h:61:53: warning: format '%lld' expects argument of type 'long long int', but argument 7 has type 'zend_long' {aka 'long int'} [-Wformat=]
       61 |   if (sp_debug_stderr > 0) dprintf(sp_debug_stderr, "[snuffleupagus][DEBUG] %s(): " fmt "\n", __FUNCTION__, ##__VA_ARGS__);
          |                                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    src/sp_upload_validation.c:48:7: note: in expansion of macro 'sp_log_debug'
       48 |       sp_log_debug("Filename: %s\nTmpname: %s\nSize: %zd\nError: %lld\nScript: %s",
          |       ^~~~~~~~~~~~

+++ src/sp_upload_validation.c
@@ -47,7 +47,7 @@ static int sp_rfc1867_callback(unsigned int event, void *event_data, void **extr
 
       sp_log_debug("Filename: %s\nTmpname: %s\nSize: %zd\nError: %lld\nScript: %s",
                    filename, tmp_name, filesize,
-                   Z_LVAL_P(zend_hash_str_find(Z_ARRVAL_P(file), "error", 5)),
+                   (unsigned long long) Z_LVAL_P(zend_hash_str_find(Z_ARRVAL_P(file), "error", 5)),
                    ZSTR_VAL(config_upload->script));
 
       cmd[0] = ZSTR_VAL(config_upload->script);
